Move Docker exceptions to Docker operator
